requirements.txt
python==3.10
pytorch==1.13.1
torchvision==0.14.1
tensorboard==2.10.0
tqdm==4.64.1
easydict==1.10
h5py==3.7.0
cuda==11.7